unify neopixel sim
