--- libs/core/sim/state.ts
@@ -1,11 +1,0 @@
-namespace pxsim {
-    export interface LightBoard extends CommonBoard {
-        // Do not laze allocate state
-        tryGetNeopixelState(pinId: number): CommonNeoPixelState;
-        neopixelState(pinId: number): CommonNeoPixelState;
-    }
-
-    export function neopixelState(pinId: number) {
-        return (board() as LightBoard).neopixelState(pinId);
-    }
-}
